+++ shared/src/commonMain/kotlin/com/ekdorn/classicdungeon/shared/ui/LayoutUI.kt
@@ -3,11 +3,11 @@ package com.ekdorn.classicdungeon.shared.ui
 internal class LayoutUI: WidgetUI() {
     val children = mutableListOf<WidgetUI>()
 
-    override fun update(elapsed: Double) {
+    override fun update (elapsed: Float) {
         children.forEach { if (it.exists && it.visible) it.update(elapsed) }
     }
 
-    override fun draw() {
+    override fun draw () {
         children.forEach { if (it.exists && it.visible) it.draw() }
     }
 }
